app/requirements.txt
asyncio
vkbottle
openai
httpx
psycopg2-binary
datetime